# skills/chapter_summarizer.py
# Chapter Summarizer Skill

import asyncio
from typing import List
from pydantic import BaseModel


class SummaryRequest(BaseModel):
    content: str
    context: str = None


class SummaryResponse(BaseModel):
    summary: str
    key_points: List[str]
    context: str = None


class ChapterSummarizer:
    """
    Skill to generate summaries of book content
    """
    
    def __init__(self):
        pass
        
    async def generate_summary(self, content: str, context: str = None) -> SummaryResponse:
        """
        Generates a concise summary of the provided content
        """
        # In a real implementation, we would use AI to summarize the content
        # For now, returning placeholder data
        summary = f"This is a summary of the chapter: {content[:100]}..."
        key_points = [
            "Key point extracted from content",
            "Another key concept",
            "Important takeaway"
        ]
        
        return SummaryResponse(
            summary=summary,
            key_points=key_points,
            context=context or "unknown"
        )