# skills/glossary_generator.py
# Glossary Generator Skill

import asyncio
from typing import List, Dict, Any
from pydantic import BaseModel


class TermDefinition(BaseModel):
    term: str
    definition: str


class GlossaryRequest(BaseModel):
    content: str
    context: str = None


class GlossaryResponse(BaseModel):
    terms: List[TermDefinition]
    context: str = None


class GlossaryGenerator:
    """
    Skill to generate glossaries of important terms from book content
    """
    
    def __init__(self):
        pass
        
    async def generate_glossary(self, content: str, context: str = None) -> GlossaryResponse:
        """
        Generates a glossary of important terms from the provided content
        """
        # In a real implementation, we would use AI to extract and define terms
        # For now, returning placeholder data
        terms = [
            TermDefinition(term="AI-native system", definition="A system where artificial intelligence is fundamentally integrated into the core architecture from the ground up."),
            TermDefinition(term="RAG", definition="Retrieval-Augmented Generation, a technique that combines retrieval from a knowledge base with language model generation."),
            TermDefinition(term="Vector Database", definition="A database that stores and retrieves data based on vector embeddings for semantic similarity.")
        ]
        
        return GlossaryResponse(
            terms=terms,
            context=context or "unknown"
        )